Reverted generic Block statement to fix the issue now and we will reintroduce the Block Trait as a separate PR afterwards.
